Fixed problem with add in CourseCollectionRepository and StudentCollectionRepository resulting in no data in DAO elements. Fixed problem with Iterator and replaced them with for-loops. Problem with system crash when student is deleted remains unsolved. Further tests required.

+++ src/main/java/se/lexicon/course_manager/data/service/student/StudentManager.java
@@ -2,6 +2,7 @@ package se.lexicon.course_manager.data.service.student;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import se.lexicon.course_manager.data.dao.CourseCollectionRepository;
 import se.lexicon.course_manager.data.dao.CourseDao;
 import se.lexicon.course_manager.data.dao.StudentDao;
 import se.lexicon.course_manager.data.service.converter.Converters;
@@ -12,6 +13,9 @@ import se.lexicon.course_manager.model.Course;
 import se.lexicon.course_manager.model.Student;
 
 
+import javax.swing.text.html.HTMLDocument;
+import java.util.Collection;
+import java.util.Iterator;
 import java.util.List;
 
 @Service
@@ -31,7 +35,9 @@ public class StudentManager implements StudentService {
     @Override
     public StudentView create(CreateStudentForm form) {
 
-        return converters.studentToStudentView(studentDao.createStudent(form.getName(), form.getEmail(),form.getAddress()));
+        Student localStudent = studentDao.createStudent(form.getName(), form.getEmail(),form.getAddress());
+
+        return converters.studentToStudentView(localStudent);
     }
 
     @Override
@@ -72,10 +78,17 @@ public class StudentManager implements StudentService {
     @Override
     public boolean deleteStudent(int id) {
 
+        if (studentDao.findById(id) == null) {
+            return false;
+        }
         Student localStudent = studentDao.findById(id);
 
-        for (Course course : courseDao.findByStudentId(id)) {
-            course.unrollStudent(localStudent);
+        if (courseDao.findByStudentId(localStudent.getId()) != null) {
+
+            for (Course course : courseDao.findByStudentId(localStudent.getId())) {
+
+                course.unrollStudent(localStudent);
+            }
         }
 
         return studentDao.removeStudent(localStudent);
